producer/requirements.txt
pandas==1.3.4
pyspark==3.1.2
scikit-learn==1.0.0
kafka-python==2.0.2
requests==2.26.0
mysql-connector-python==8.0.26
pymongo==3.12.0

beautifulsoup4==4.7.1
lxml==4.6.3
uvicorn==0.14.0
fastapi==0.68.0
sweetviz==2.1.3
